Release v1.2.0

+++ src/zeal_cli.py
@@ -1,5 +1,6 @@
 import argparse
 import shutil
+import sys
 
 import zeal
 
@@ -12,10 +13,19 @@ def main():
         "install", help="Install one or more docsets. See `zeal-cli install --help`"
     )
     install_command.add_argument(
-        "docsets", nargs="*", help="A list of docset names, separated by a space."
+        "docsets", nargs="*", help=(
+            "A list of docset names, separated by a space. "
+            "A specific version of a docset can be selected for installation by following "
+            "the docset name with an equals (=) and the version of the docset to select"
+        )
     )
 
     subparsers.add_parser("list", help="Prints a list of installed docsets")
+
+    search_command = subparsers.add_parser("search", help="Prints a list of installed docsets")
+    search_command.add_argument(
+        "docset", nargs=1, help="A name of a docset."
+    )
 
     remove_command = subparsers.add_parser(
         "remove", help="Delete one or more docsets. See `zeal-cli remove --help`"
@@ -50,15 +60,29 @@ def main():
         "value", help="The value to assign to the specified config item."
     )
 
+    parser.add_argument(
+        "-v", "--version", action="store_true", help="Print Zeal-CLI's version information"
+    )
+
     args = parser.parse_args()
 
+    if args.version:
+        print(f"Build Version: {zeal.version.build_version}")
+        sys.exit()
     if args.action == "install":
         if args.docsets:
             print("Getting list of available docsets")
             feeds = zeal.downloads.get_feeds()
             for docset in args.docsets:
+                docset_name = docset
+                docset_version = zeal.docset.LATEST_VERSION
+                # Parse version string if defined
+                if '=' in docset:
+                    docset_info = docset.split('=', 1)
+                    docset_name = docset_info[0]
+                    docset_version = docset_info[1]
                 print(f"Installing docset: {docset}")
-                zeal.docset.download(docset, feeds)
+                zeal.docset.download(docset_name, feeds, docset_version=docset_version)
                 print(f"Successfully installed docset: {docset}")
             print("Cleaning up")
             shutil.rmtree(feeds)
@@ -67,7 +91,13 @@ def main():
             install_command.print_help()
 
     elif args.action == "list":
-        print(*zeal.docset.list_all(), sep="\n")
+            print(*zeal.docset.list_all(), sep="\n")
+
+    elif args.action == "search":
+        feeds = zeal.downloads.get_feeds()
+        docset_name = args.docset[0]
+        docset_versions = zeal.docset.get_docset_versions(docset_name, feeds)
+        print(f"Available versions for docset {docset_name}: {', '.join(docset_versions)}")
 
     elif args.action == "remove":
         if args.docsets:
